feat(navigation): add NavigationActions and the composable SideBarNavigation (#19)
* feat: implement the NavigationActions.kt and SideBarNavigation.kt

Implement the different routes and screens necessary for the navigation
through the app. The SideBarNavigations serves as a Menu which will be
available in the main Screens.

* feat: implement the NavigationActions.kt and SideBarNavigation.kt

Implement the different routes and screens necessary for the navigation
through the app. The SideBarNavigations serves as a Menu which will be
available in the main Screens.

* feat: implement NavigationActionsTest.kt

Start implementing the tests for the different methods of
NavigationsActions.

* feat: implement SideBarNavigationTest.kt with TestTags, update the gradle.

Implement the tests for SideBarNavigation with some tweaks to the code
to make the test tags available. changed gradle build to include mockito
imports.

* feat: add documentation to both NavigationActions.kt and SideBarNavigation.kt

* fix: run ktfmtFormat

* feat: add tests to NavigationActionsTest.kt and SideBarNavigationTest.kt

Add tests to both test files. Also changed the icons in the
SideBarNavigation to make it more visually apealing.

* feat: add tests to SideBarNavigationTests.kt

* fix: bug in test in SideBarNavigationTest.kt drawer should be open even if list is empty

* feat: add tests to SideBarNavigationTest.kt

* feat: add tests to SideBarNavigationTests.kt and NavigationActionsTests.kt

Add tests to both test classes, specially for the NavOptions properties
which were previously inaccesible through the NavController.

* fix: dependency issues in tests, removed Auth as a TopLevelDestination, update the JVM version

* fix: remove comment GradelDependency

* fix: remove unnecessary import in SideBarNavigation.kt

* fix: change according to PR review.

* fix: make AUTH a TopLevelDestination