feat(map): load routes from view model
Routes displayed in the hikes list of the map are now loaded from the view model. Every time the user moves the map, the displayed area is updated in the view model to update the hikes as well.

Still need to have a cooldown for API requests, since the updates can be quite frequent right now.