fix(map): only initialize list of hikes once
The previous implementation of the map composable, in particular the CollapsibleHikesList composable, would have refreshed the list of hikes on every recompose, which goes against good practices and performance optimization.

The map composable will now only refresh the hikes on the first composing, then further re-compositions won't trigger any update of the hiking routes. The only way those routes get updated is when the user interacts with the map.