feat(map): implement hikes list item composable
Instead of a simple Text composable, hiking routes are now displayed each in their own HikingRouteItem composable, which was designed to resemble the Figma as much as possible.

There is no elevation plot for now as the elevation data is not available. A placeholder was put there instead.

The background color of the expandable hiking routes list was also changed to better fit the Figma mockup.