refactor(map): remove unnecessary todos
Removed a TODO about setting the initial map zoom. The current value looks good for now and will be changed if need be in the future.

As for the initial map center, it is hardcoded now but should be the user's location in the future, when the location permission has been implemented.