feat(map): hide button while searching
To avoid the user spamming the search button, we provide UI feedback by hiding the button while the search is performed. The button appears again once the search has been completed, either having succeeded or failed.

For this, I added callbacks to the methods of the view model, because the screen needs to be called when the request is finished, I believe this is the cleanest way of doing it.