test(map): add more specific tests to map
Now that the repository and the view model are merged, the repository can be mocked to test the reactions of the map when user input is simulated. This gives the opportunity of implementing better and more robust tests than before.

Also added tests to check the new search button behavior (show/hide search button, show a searching message and animation while the search is ongoing, and so on).