test(map): fix clickingOnHikeItemSelectsHikeRoute
The test failed because when it was about to click on an item to select a route, it did not find the UI element representing the list item. None were displayed.

The solution to this was to add a `waitForIdle()` call to wait for recomposition to finish before performing the click.