feat(saved-hikes): create SavedHike data type
To implement the saved hikes feature, one needs to start by thinking how to represent those saved hikes. Features might be added later, but for now, the most important parts are the id of the hike to retrieve additional information about it, its name to be displayed in the UI and a date at which the user plans on going on this hike.