refactor(saved-hikes): rename repository to dummy
The dummy repository that was implemented in this PR is kind of a dummy placeholder, in the sense that it does not do anything except providing fake data.

Hence, it should not be have Firebase in its name, but rather be called RepositoryDummy and the Firebase one will be implemented later and switched in the dummy repository's place, as pointed out in the review.