refactor: add message params to HikeCard
To make the HikeCard composable more generic to use it somewhere else than the map, we need to add some parameters to have other possible messages, not just "suitable" and "challenging". This is done in this commit by adding three parameters instead of the old `isSuitable`.

- messageIcon defines the icon to display to the left of the message
- messageContent defines the text of the message
- messageColor defines the tint of both the text and the icon

This way, the message displayed below each hike card is more customizable.

If any of the previously mentioned parameters is null, the message is not displayed.