feat(nav): pass scaffold padding as content param (#76)
The sidebar component adds a burger menu button in the top-left corner of the app. For the map screen, it is no problem if this button hovers a little bit on the map, the user can simply move the map to see what was below the button.

In other screens however, titles might get hidden behind the menu button. Adding a hardcoded padding is not a good solution since the positioning and padding of the menu button will probably change in the future.

However, the scaffold allows us to use its top padding to know exactly how much space the menu button is taking. The composable that uses the menu then may or may not use the padding, depending on the use case.

Those changes are retro-compatible and tests should pass, because Kotlin allows a parameter to be ignored in lambda functions.