refactor(saved-hikes): remove nearby for now
Instead of having a placeholder "Coming Soon" screen for nearby hikes, I think it would be better to add it once we actually decide to implement it.

From internal discussion, the initial scope we thought about for the nearby hikes section might be infeasible or take too much time. We have to decide before adding an actual nearby section, even "Coming Soon".