feat(util): implement humanReadablePlannedLabel
For the saved hikes screen, I needed a function that converts a date into a human readable label, such as "Planned for today" or "Planned for tomorrow", or "Planned for next week", you get the idea.

This turned out to be surprisingly nerve-wrecking, as I had it done in a localizable manner, and had to test it too (tests are in the next commit).