refactor: HikeCard now takes a painter as param
The HikeCard composable was taking an ImageVector as a parameter, but image vectors are limited and sometimes require additional libraries to be available.

Using local resources ensures that the SVG is available and is the one we want exactly, it also allows us to take one from the Internet or design it by ourselves if needed.