fix(map): fix search crashing when zoomed out
When the user zoomed out too much of the map, horizontal and vertical repetitions were previously enabled, which means the world map was repeated horizontally and vertically when the user reached the edge.

This lead to the bounds being invalid : the map started at a given longitude on one map, and finished on another, smaller longitude on the next map repetition, which lead to the max longitude being smaller than the min longitude, leading to the app crashing.

Disabling the horizontal and vertical repetition of the map is a clean way to avoid this without explicitly checking for the bounds to be valid, which would be like hardcoding a fix.