fix: lock orientation to remove landscape mode (#86)
The app is not adapted to landscape mode, at least yet, and we do not plan to support it as of now.

The chosen fix is to lock the orientation in `MainActivity` directly. Because Jetpack Compose uses a single activity, this locks the orientation for the whole app.

This way of fixing was chosen instead of another possible approach in `AndroidManifest.xml`, because that is part of a discouraged API. Locking the orientation in `MainActivity` is a better alternative because it is not discouraged (at least it does not generate a `DiscouragedApi` warning).

No test was added for this lock. The only way of testing orientation I could find was to set the requested orientation of the activity, which overrides the lock I put in `MainActivity`, hence the test will always fail because the app will indeed go into landscape mode, even though an actual user would not be able to do this.