fix(map): set user agent before OSMDroid map init (#89)
We use OSMDroid to display map tiles from OSM as a map view. OSMDroid requires to set an actual user agent for the tiles requests.

The configured user agent was indeed set to a valid value in our code, but it was set after the map view was created, meaning the first requests were being sent before the configuration was set, and an error could be seen in Logcat. It did not impact the functionality of the app too much, but is better fixed anyway.