feat(saved-hikes): implement saved hikes screen
See #77.

* refactor(map): extract list item composable

Since the saved hikes will require a composable similar to what already exists on the map, the existing one was extracted from the MapScreen.kt file to become a component of its own, usable by the MapScreen but also another screen if needed.

* refactor: add message params to HikeCard

To make the HikeCard composable more generic to use it somewhere else than the map, we need to add some parameters to have other possible messages, not just "suitable" and "challenging". This is done in this commit by adding three parameters instead of the old `isSuitable`.

- messageIcon defines the icon to display to the left of the message
- messageContent defines the text of the message
- messageColor defines the tint of both the text and the icon

This way, the message displayed below each hike card is more customizable.

If any of the previously mentioned parameters is null, the message is not displayed.

* refactor(map): reduce cognitive complexity

Sonar was not happy about the CollapsibleHikesList composable's cognitive complexity, so I extracted the logic to create a new HikeCard and made it another composable.

* feat(saved-hikes): create saved hikes screen

Created a composable that will contain the saved hikes screen once it is implemented. Added it to the definition of navigation in MainActivity.kt.

* refactor(nav): rename "planned" hikes to "saved"

The saved hikes will not only contain "planned" hikes but also hikes the user has saved without scheduling them for any date in the future yet, hence "saved" hikes sounds like a more appropriate name.

* feat(saved-hikes): add bottom menu navigation

Started implementing the Saved Hikes screen by creating its skeleton with the bottom navigation menu and the nearby, planned and saved sections.

Each section contains a placeholder text for now.

* feat(saved-hikes): implement saved hikes section

The saved hikes section of the Saved Hikes screen is one of the three available sections. It lists the hikes the user saved without a planned date.

To this end, the SavedHikesScreen needed to take an additional parameter, the view model for the saved hikes.

* feat(util): timestamp from year, month and day

Implemented a util function that creates a Firebase Timestamp from a year (Int), a month (Int) and a day (Int). Also implemented tests for that function.

* feat(saved-hikes): add more dummy saved hikes

To test the Saved Hikes screen properly, there needs to be some saved hikes with a planned date. I add some to SavedHikesRepositoryDummy using the newly created util function that creates a timestamp from a year, month and day.

* refactor(saved-hikes): remove nearby for now

Instead of having a placeholder "Coming Soon" screen for nearby hikes, I think it would be better to add it once we actually decide to implement it.

From internal discussion, the initial scope we thought about for the nearby hikes section might be infeasible or take too much time. We have to decide before adding an actual nearby section, even "Coming Soon".

* fix(saved-hikes): fix menu hiding section title

The sidebar menu button was hiding the saved hikes screen section title ("Saved hikes" or "Planned hikes") because of how the sidebar was implemented.

I fixed the sidebar implementation in another PR, merged it into main, merged main into this branch, and now I can finally fix this lol.

* fix(saved-hikes): localize saved hikes title

Made the title of the saved hikes section in the saved hikes screen localized, so that it can be easily translated.

Basically, this means I moved a hardcoded string into the `strings.xml`.

* feat(util): implement humanReadablePlannedLabel

For the saved hikes screen, I needed a function that converts a date into a human readable label, such as "Planned for today" or "Planned for tomorrow", or "Planned for next week", you get the idea.

This turned out to be surprisingly nerve-wrecking, as I had it done in a localizable manner, and had to test it too (tests are in the next commit).

* test(util): tests for humanReadablePlannedLabel

Tests for the humanReadablePlannedLabel util function. Those tests turned out to be quite difficult to write, but I opted for a parametric approach that allowed me to simply write some test data and watch the tests get performed in loops. Very satisfying.

* feat(saved-hikes): add planned hikes section

Added a new (second) section in the Saved Hikes screen that lists only saved hikes with a date associated to it (i.e. the user planned them for some date).

The date is displayed below the corresponding hike card in a human-readable format, for example "Planned next week (1 November 2024)".

* refactor: HikeCard now takes a painter as param

The HikeCard composable was taking an ImageVector as a parameter, but image vectors are limited and sometimes require additional libraries to be available.

Using local resources ensures that the SVG is available and is the one we want exactly, it also allows us to take one from the Internet or design it by ourselves if needed.

* feat(saved-hikes): no hikes to display message

If there are no saved hikes or planned hikes to be displayed in the respective section, a message will be displayed instead, such as "No planned hikes to display".

* feat(saved-hikes): placeholder for onClick event

Instead of doing nothing when a saved hike or planned hike is clicked, display a toast telling the user that this feature is not ready yet, to avoid confusion.

* test(saved-hikes): add integration tests

Added integration (UI) tests for the Saved Hikes screen.

* test(saved-hikes): adapt previous tests

Because the Saved Hikes screen was implemented, the placeholder screen is no more, and hold tests were failing because they assumed the Saved Hikes screen was a placeholder.

I fixed those, they are now asserting correctly about the new Saved Hikes screen.

* refactor: remove unused vector resource

Sonar pointed out that this resource is not used anywhere, which is right, because I removed the nearby hikes section of the saved hikes screen for now, so the icon is nowhere to be found in the app.

We'll add it back when we need it.

* fix(saved-hikes): use hike name instead of hike id

Realized the SavedHike type has a name attribute that should be used instead of the ID when displaying the hikes.

* fix: made altitude texts in HikeCard localizable

As pointed out in the PR review, this text should have been in strings.xml from the beginning. Better late that never.

* test(utils): add datetime tests for edge cases

Added tests for the DatetimeUtils.kt file. Tested Timestamp.from(year, month, day) with invalid years, months, and days, and with borderline years, months, and days.

What I mean by borderline is that if 0 is an invalid value and 1 is the first valid value for a year, then 1 is a borderline value.