test(map): use dispatcher instead of wait clauses
The previous version of the map screen tests used `composeTestRule.waitUntil` and `Thread.sleep` clauses, which works, but is not that clean nor reliable.

Using an UnconfinedTestDispatcher in the view model runs all view model coroutines on the main thread, hence the test will naturally block until the view model's operations are finished, and will only check the UI afterwards.

It does not, however, bring a significant change in the time the tests take to get executed. Tests do not take longer, but they're not faster either.