feat(map): display hike names instead of ids (#92)
In the map screen's list of hikes, display each hike's name instead of its ID.

Added tests to check the behavior when the hike has a name or has no name (display a default name in this case).

Refactored the tests of the map screen to make them cleaner and reliable using a test dispatcher for coroutines instead of wait clauses on the composeTestRule. However, there was no significant change in how much time the tests take to terminate.