feat(map): add user puck, center on user location
Implemented the location features on the map screen directly.

Since permissions have to be asked for in the UI directly, a view model or a model do not make much sense to handle the whole. In my next commits, however, I will aim at separating the logic into multiple components and divide the functionality, rather than having a big block in MapScreen.

There is still a lot of work, but this POC works.