feat(perms): implement first permission utils
Permission requests in Android are VERY convoluted (at least in my opinion).

You want to request location permission. Good. You launch a permission request. Yes but it's not that simple. If you simply launch it, it might never appear if you already launched two and the user denied each time.

Hence you need to check. Android provides the variable `shouldShowRationale` to indicate whether you should show an explanation dialog to the user about why you need this permission.

If `shouldShowRationale` is true, you may launch a request safely, but if it's false, either the user never saw any request from you for this permission (in this case you can launch it safely), or the user completely denied it and you may never launch a request in that case.

To differentiate between those two cases, Android can't tell you whether you have already asked the user, so you have to have a SharedPreferences instance to store whether you asked that permission already.

It is a nightmare but it's going to work at some point hopefully maybe.