fix(map): fix bad disposal of the map view
When the location was added, a bug was introduced. The map asks for periodic updates of the user location, but does not stop the periodic updates once the screen is left.

The periodic update is then run in the background and tries to update the marker's position on the map view, but the map view has been left and its repository has been set to a null pointer, leading to a NullPointerException.

This fix prevents the crash from happening and disposes correctly of the map view.