feat(map): center map once permission was granted
In the previous versions, the user had to click a second time on the "Center on my location" button to actually center the map, the first press only asked for the permission but did not do anything once the permission was granted.

Now the app keeps track of whether the button was pressed and centers the map on the user's location if it was.