feat(hikes): add selectRouteById (#141)
When clicking on a route in the saved hikes screen, we want to open the details page of the hike but don't have the corresponding HikeRoute instance. This method allows to select the hike nevertheless using only its ID, by retrieving the data from the repository.

Added tests as well.