feat(saved): better error handling (#151)
Improves the error handling in the saved hikes screen. Displays a loading animation while saved hikes are being loaded. If an error occurs, an error message is displayed with a refresh button, allowing the user to try again.

This also adds two generic components : a centered loading animation, and a centered error message with an optional icon button for action.

See #151 for more details.