refactor: replace and remove toFormattedDate (#218)
This function used a date format that did not adapt to the current Locale. It used the dd/MM/yyyy format, but this format can be mistaken for MM/dd/yyyy depending on the date, which can be confusing for English speakers who are used to the second option rather than the first.