fix(hikes): fix loading and OSM data available status
Loading was not being updated correctly.

Say you have loading operation 1 that starts and thus sets _loading.value to true.

Then loading operation 2 starts and also sets _loading.value to true, which has no effect.

Then loading operation 1 terminates and sets _loading.value to false, because it ended, but it's wrong because loading operation 2 is still ongoing.

A while after, loading operation may terminate too, but _loading.value is already false.

I implemented a counter to keep track of how many ongoing loading operations there are and update _loading accordingly.

Also, allOsmDataLoaded was not updated correctly in retrieveLoadedHikesOsmData, I fixed it by calling an update helper explicitly.

retrieveLoadedHikesOsmData now has a mutex that prevents multiple requests to be sent to OSM at the same time. If two calls are made, only one will run at a time. This allows for the hikes to be updated in the meantime and for the second call to realise it's not needed and to return early.