fix(map): prevent displaying saved hikes
If the user was on the saved hikes screen and saved hikes have been loaded into the view model, then the user goes back to the map screen, the saved hikes would be displayed on the map and in the list.

This is not desired behavior, at least for now, so I fixed it by using the newly exposed HikesViewModel.loadedHikesType.