feat(hikes): new HikesViewModel flows (#232)
This commit introduces a few improvements in the HikesViewModel implementation:

- `allOsmDataLoaded` is a new public state flow of the view model that indicates whether all hikes in `hikeFlows` have their OSM data available
- `loading` is now handled in a better way, and only set to false when all loading operations are finished
- `loadedHikesType` indicates what type of hikes are loaded in `hikeFlows`, whether they are saved hikes, hikes loaded from OSM bounds, or just none at all

Also, `retrieveLoadedHikesOsmData` has been updated to only allow for one request at a time (using a Mutex). This way, only one request to get OSM data for hikes is being done at once, and the next requests might just never be sent if they are not needed (if all the hikes' OSM data are already available).