fix(run): fix run this hike button disabling
Small bug fix : by going very quickly between the `HikeDetailsScreen` and the `RunHikeScreen` by clicking on the *Run this hike* button and *Stop the run* over and over, the *Run this hike* button ended up "disabling" in the sense that clicking it did not do anything.

This was because `wantToRunHike` did not switch back to `false` if we clicked fast enough. This is now fixed.

Co-authored-by: Mathéo Chacón Ducommun-Dit-Verron <151830561+mchac15@users.noreply.github.com>