fix(hikes): unsaving hike in details (#304)
There was a bug when the user went to the saved screen and opened a hike's details. If the user then unsaved the hike, it removed the hike from the loaded hikes, thus unselected then hike, and thus kicked the user out of the details screen.

This bug was "by design" in the sense that the HikesViewModel was built with the idea that when the selected hike is unloaded, it must be unselected as well. But this does not make sense from a practical point of view.

This commit corrects the design of HikesViewModel and its tests to let the user unsave the hike freely.