fix: display planned hikes dates (#312)
The display of the planned date for planned hikes was mistakenly removed in a previous PR. This commit introduces it again with a regression test.

HikeCard was performing a check for elevation and did not display the message, even if it was provided, before the elevation was set.

This commit also removes the check on elevation to make sure the message is displayed independently of the elevation graph.